+++ src/index.js
@@ -11,6 +11,8 @@ export * from './elementals/menu/index.js';
 export * from './elementals/modal/index.js';
 export * from './elementals/navbar/index.js';
 export * from './elementals/segmented/index.js';
+export * from './elementals/suggest/index.js';
 export * from './elementals/switch/index.js';
 export * from './elementals/tabs/index.js';
+export * from './elementals/toolbar/index.js';
 export * from './elementals/tooltip/index.js';